% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decomp.R
\name{bw_decomp}
\alias{bw_decomp}
\title{between-within decomposition of lifespan inequality measures}
\usage{
bw_decomp(
  age,
  ax,
  dx,
  lx,
  ex,
  prop,
  method = c("Theil", "edag", "variance", "MLD", "Gini")
)
}
\arguments{
\item{age}{numeric. vector of lower age bounds.}

\item{ax}{numeric. vector of the average time spent in the age interval of those dying within the interval.}

\item{dx}{numeric. vector of the lifetable death distribution.}

\item{lx}{numeric. vector of the lifetable survivorship.}

\item{ex}{numeric. vector of remaining life expectancy.}

\item{method}{character one of \code{"Theil", "edag","variance","MLD","Gini"}}
}
\description{
Partition a lifespan inequality index into additive components of between-group inequality and within-group inequality. Presently implemented for Theil's index, e-edagger, variance, mean log deviation, and the gini coeficient.
}
